+++ src/Http/Controllers/MappingController.php
@@ -95,7 +95,7 @@ class MappingController extends Controller
 
         if (is_null($shopifyMapping)) {
 
-            session()->flash('error', 'Export mapping save failed. Please run the Shopify install command.');
+            session()->flash('error', trans('shopify::app.shopify.export.mapping.save_failed'));
 
             return redirect()->back();
         }