+++ tests/Feature/Validator/ShopifyProductValidatorTest.php
@@ -7,7 +7,7 @@ beforeEach(function () {
     $this->validator = new ShopifyProductValidator();
 });
 
-it('should passes validation with valid data', function () {
+it('should pass validation with valid data', function () {
     $data = [
         'filters' => [
             'credentials' => 1,
@@ -20,7 +20,7 @@ it('should passes validation with valid data', function () {
     expect($validator->passes())->toBeTrue();
 });
 
-it('should fails when credentials are missing', function () {
+it('should fail when credentials are missing', function () {
     $data = [
         'filters' => [
             'channel'  => 'shopify_default',
@@ -33,7 +33,7 @@ it('should fails when credentials are missing', function () {
         ->and($validator->errors()->keys())->toContain('filters.credentials');
 });
 
-it('should fails when channel is missing', function () {
+it('should fail when channel is missing', function () {
     $data = [
         'filters' => [
             'credentials' => 1,
@@ -46,7 +46,7 @@ it('should fails when channel is missing', function () {
         ->and($validator->errors()->keys())->toContain('filters.channel');
 });
 
-it('should fails when currency is missing', function () {
+it('should fail when currency is missing', function () {
     $data = [
         'filters' => [
             'credentials' => 1,
@@ -59,7 +59,7 @@ it('should fails when currency is missing', function () {
         ->and($validator->errors()->keys())->toContain('filters.currency');
 });
 
-it('should fails when credentials is not integer', function () {
+it('should fail when credentials is not integer', function () {
     $data = [
         'filters' => [
             'credentials' => 'abc',